+++ src/TestEngine/nunit-agent/Program.cs
@@ -31,7 +31,7 @@ using NUnit.Engine;
 using NUnit.Engine.Agents;
 using NUnit.Engine.Internal;
 using NUnit.Engine.Services;
-using NUnit.Engine.Utilities;
+using NUnit.Engine.Helpers;
 
 namespace NUnit.Agent
 {
