+++ src/TestEngine/nunit.engine.core/Agents/RemoteTestAgent.cs
@@ -27,7 +27,7 @@ using System.Runtime.Remoting.Channels;
 using System.Runtime.Remoting.Channels.Tcp;
 using System.Threading;
 using NUnit.Engine.Internal;
-using NUnit.Engine.Utilities;
+using NUnit.Engine.Helpers;
 
 namespace NUnit.Engine.Agents
 {
