+++ src/TestEngine/nunit.engine.core/Helpers/CecilExtensions.cs
@@ -1,5 +1,5 @@
 // ***********************************************************************
-// Copyright (c) 2019 Charlie Poole, Rob Prouse
+// Copyright (c) 2015 Charlie Poole, Rob Prouse
 //
 // Permission is hereby granted, free of charge, to any person obtaining
 // a copy of this software and associated documentation files (the
@@ -8,10 +8,10 @@
 // distribute, sublicense, and/or sell copies of the Software, and to
 // permit persons to whom the Software is furnished to do so, subject to
 // the following conditions:
-//
+// 
 // The above copyright notice and this permission notice shall be
 // included in all copies or substantial portions of the Software.
-//
+// 
 // THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
 // EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
 // MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
@@ -23,48 +23,46 @@
 
 using System;
 using System.Collections.Generic;
-using NUnit.Engine.Utilities;
+using Mono.Cecil;
 
-namespace NUnit.Engine.Extensions
+namespace NUnit.Engine.Helpers
 {
-    public delegate bool TestPackageSelectorDelegate(TestPackage p);
-
     /// <summary>
-    /// Extension methods for use with TestPackages
+    /// Extension methods that make it easier to work with Mono.Cecil.
     /// </summary>
-    public static class TestPackageExtensions
+    internal static class CecilExtensions
     {
-        public static bool IsAssemblyPackage(this TestPackage package)
+        public static List<CustomAttribute> GetAttributes(this TypeDefinition type, string fullName)
         {
-            return package.FullName != null && PathUtils.IsAssemblyFileType(package.FullName);
+            var attributes = new List<CustomAttribute>();
+
+            foreach (CustomAttribute attr in type.CustomAttributes)
+            {
+                if (attr.AttributeType.FullName == fullName)
+                    attributes.Add(attr);
+            }
+
+            return attributes;
         }
 
-        public static IList<TestPackage> AssemblyPackages(this TestPackage package)
+        public static CustomAttribute GetAttribute(this TypeDefinition type, string fullName)
         {
-            return package.Select(p => p.IsAssemblyPackage());
+            foreach (CustomAttribute attr in type.CustomAttributes)
+            {
+                if (attr.AttributeType.FullName == fullName)
+                    return attr;
+            }
+
+            return null;
         }
 
-        public static bool HasSubPackages(this TestPackage package)
+        public static object GetNamedArgument(this CustomAttribute attr, string name)
         {
-            return package.SubPackages.Count > 0;
-        }
+            foreach (var property in attr.Properties)
+                if (property.Name == name)
+                    return property.Argument.Value;
 
-        public static IList<TestPackage> Select(this TestPackage package, TestPackageSelectorDelegate selector)
-        {
-            var selection = new List<TestPackage>();
-
-            AccumulatePackages(package, selection, selector);
-
-            return selection;
-        }
-
-        private static void AccumulatePackages(TestPackage package, IList<TestPackage> selection, TestPackageSelectorDelegate selector)
-        {
-            if (selector(package))
-                selection.Add(package);
-
-            foreach (var subPackage in package.SubPackages)
-                AccumulatePackages(subPackage, selection, selector);
+            return null;
         }
     }
 }
