+++ src/TestEngine/nunit.engine.core/Helpers/PathUtils.cs
@@ -26,7 +26,7 @@ using System.IO;
 using System.Text;
 using System.Collections.Generic;
 
-namespace NUnit.Engine.Utilities
+namespace NUnit.Engine.Helpers
 {
     /// <summary>
     /// Static methods for manipulating project paths, including both directories
