+++ src/TestEngine/nunit.engine.core/Helpers/ResultHelper.cs
@@ -27,7 +27,7 @@ using System.Globalization;
 using System.Reflection;
 using System.Xml;
 
-namespace NUnit.Engine.Utilities
+namespace NUnit.Engine.Helpers
 {
     /// <summary>
     /// ResultHelper provides static methods for working with
