+++ src/TestEngine/nunit.engine.tests/Services/ResultWriters/XmlOutputTest.cs
@@ -26,7 +26,7 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Reflection;
-using NUnit.Engine.Utilities;
+using NUnit.Engine.Helpers;
 
 namespace NUnit.Engine.Services.ResultWriters.Tests
 {
