+++ src/TestEngine/nunit.engine.tests/Utilities/AssemblyHelperTests.cs
@@ -25,7 +25,7 @@
 using System.IO;
 using NUnit.Framework;
 
-namespace NUnit.Engine.Utilities
+namespace NUnit.Engine.Helpers
 {
     [TestFixture]
     public class AssemblyHelperTests
