+++ src/TestEngine/nunit.engine.tests/Utilities/ProcessUtilsTests.cs
@@ -24,7 +24,7 @@
 using NUnit.Framework;
 using System.Text;
 
-namespace NUnit.Engine.Utilities
+namespace NUnit.Engine.Helpers
 {
     public static class ProcessUtilsTests
     {
