+++ src/TestEngine/nunit.engine/Services/RuntimeFrameworkService.cs
@@ -28,7 +28,7 @@ using System.IO;
 using Mono.Cecil;
 using NUnit.Common;
 using NUnit.Engine.Internal;
-using NUnit.Engine.Utilities;
+using NUnit.Engine.Helpers;
 
 namespace NUnit.Engine.Services
 {
